[Refactoring] Separate presentational components
- Separate by folder components, containers, pages
