+++ src/containers/auth/LoginFormContainer.jsx
@@ -2,16 +2,16 @@ import React, { useCallback } from 'react';
 
 import { useDispatch, useSelector } from 'react-redux';
 
-import LoginForm from './LoginForm';
-import LogoutForm from './LogoutForm';
+import LoginForm from '../../components/auth/LoginForm';
+import LogoutForm from '../../components/auth/LogoutForm';
 
 import {
   changeLoginField,
   requestLogin,
   logout,
-} from './slice';
+} from '../../slice';
 
-import { get } from './utils';
+import { get } from '../../utils';
 
 export default function LoginFormContainer() {
   const dispatch = useDispatch();
