+++ src/containers/menu/CategoriesContainer.test.jsx
